+++ src/list-select/ListSelectWithLocalSearch.component.js
@@ -26,7 +26,7 @@ export default React.createClass({
     },
 
     render() {
-        const listStyle = { width: '100%', outline: 'none', border: 'none', padding: '0rem 1rem' };
+        const listStyle = { width: '100%', outline: 'none', border: 'none', padding: '0rem 1rem', overflowX: 'auto' };
 
         return (
             <div>